Tests: consolidate Booksleeve suite
Long overdue collapsing of the two suites
