+++ StackExchange.Redis.Tests/Helpers/redis-sharp.cs
