+++ StackExchange.Redis.Tests/Issues/SO10825542.cs
@@ -3,16 +3,16 @@ using System.Text;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace StackExchange.Redis.Tests.Booksleeve.Issues
+namespace StackExchange.Redis.Tests.Issues
 {
-    public class SO10825542 : BookSleeveTestBase
+    public class SO10825542 : TestBase
     {
         public SO10825542(ITestOutputHelper output) : base(output) { }
 
         [Fact]
         public void Execute()
         {
-            using (var muxer = GetUnsecuredConnection())
+            using (var muxer = Create())
             {
                 var key = Me();
